Ripped out Primus and replaced with socket.io

+++ server/rewriteGameIndex.js
@@ -1,9 +1,10 @@
 const fs = require("fs");
 const ip = require("ip");
 const path = require('path');
+const rimraf = require('rimraf');
 const rootDir = path.join(__dirname, '..');
 
-const ANIMATION_FOLDER = "hypeAnimation3";
+const ANIMATION_FOLDER = "hypeAnimation";
 const EXPORT_NAME = "nebula_derby.hyperesources";
 const BETTER_EXPORT_NAME = "nebula_derby_hyperesources";
 
@@ -29,9 +30,18 @@ function getData(fileName, type) {
 function renamer(oldName, newName) {
   console.log('Renamer');
   if (fs.existsSync(oldName)) {
-    fs.renameSync(oldName, newName);
+    if (fs.existsSync(newName)) {
+      rimraf.sync(newName, fs, () => {
+        fs.renameSync(oldName, newName);
+      });
+    } else {
+      fs.renameSync(oldName, newName);
+    }
   } else if (fs.existsSync(`${rootDir}/${ANIMATION_FOLDER}/${initHypeFile}`)) {
     renamer(`${rootDir}/${ANIMATION_FOLDER}/${initHypeFile}`, `${rootDir}/${ANIMATION_FOLDER}/${newHypeFile}`);
+  }
+  if (fs.existsSync(newName)) {
+    console.log(`Renamer complete`);
   } else {
     console.log(`Renamer target ${oldName} not found`);
   }
@@ -45,7 +55,11 @@ module.exports = function () {
       const string = data.toString('utf8');
       const replacement = `nebula_derby_hyperesources`;
       const m = string.replace(/nebula_derby.hyperesources/g, replacement);
-      fs.writeFileSync(ANIMATION_FILE, m, (err) => {
+
+      const socketReplacement = `http://localhost:3001/game/nebula_derby_hyperesources/socket.io.slim.js`;
+      const n = m.replace(/nebula_derby_hyperesources\/socket.io.slim.js/g, socketReplacement);
+
+      fs.writeFileSync(ANIMATION_FILE, n, (err) => {
         if (err) throw err;
         console.log('Replaced!');
       });
